requirements.txt
fastapi
uvicorn[standard]
rdkit
pydot
Pillow
jinja2
python-multipart
